Patch contributed by Edwin Torok: Fixed the run-time so that it uses intptr_t when casting between integers and pointers. Made other improvements for building on 64-bit platforms.
git-svn-id: http://llvm.org/svn/llvm-project/safecode/trunk@88955 91177308-0d34-0410-b5e6-96231b3b80d8

+++ runtime/DebugRuntime/DebugReport.h
@@ -34,7 +34,7 @@ struct OutOfBoundsViolation : public DebugViolationInfo {
   //  objstart - The start of the object in which the source pointer was found.
   //  objlen   - The length of the object in which the source pointer was found.
   const void * objStart;
-  unsigned int objLen;
+  ptrdiff_t objLen;
   virtual void print (std::ostream & OS) const;
 };
